# openapi.yaml
openapi: 3.1.0
info:
  title: DSA Faculty Service API
  version: 0.4.0
  description: |
    **DSA Faculty Service** — сервис агрегации и обогащения данных о научно-педагогических
    работниках (НПР) НИУ ВШЭ для платформы **Digital Student Assistant**.

    Отвечает за слой данных: сбор профилей с hse.ru и из скрытого API публикаций,
    нормализация, хранение, выдача через REST.

    **Статус:** draft v0.2. Источник истины — этот файл.

    **Изменения v0.1 → v0.2:**
    - `positions` — только должности (без `employment_addition`, `managers` вынесены в `relations`)
    - `education` — только `degrees` + `extra_education`
    - `awards`, `interests`, `grants`, `editorial_staff`, `conferences` — на верхнем уровне `Person`
    - Публикации — отдельная сущность, many-to-many через `authorships`
    - `/news` — лента последних публикаций + (позже) новостей портала
    - `/admin/scrape` — запуск парсера из API (интеграция с Jupyter-логикой)
  contact:
    name: Ivan Agafoshin
    url: https://github.com/<TBD>/dsa-faculty-service
  license:
    name: MIT
servers:
  - url: https://faculty.dsa.example.com/api/v1
    description: Production TBD (subdomain on DSA domain)
  - url: http://localhost:8000/api/v1
    description: Local dev
...
paths:
...
  /admin/scrape:
    post:
      tags: [admin]
      summary: Start a scrape job
      description: |
        Enqueues a background HSE profile scrape and returns the new `job_id` immediately.
        Poll `GET /admin/scrape/{job_id}` for progress.
      operationId: runScrape
      security: [{ adminToken: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 20000 } }
        - name: campus_ids
          in: query
          description: HSE campus IDs (1125608=Moscow, 1125609=SPb, 1125610=NN, 1125611=Perm). Empty = all.
          style: form
          explode: true
          schema:
            type: array
            items: { type: string }
        - name: letters
          in: query
          description: First letters of surnames (Cyrillic А, Б, ...). Empty = all.
          style: form
          explode: true
          schema:
            type: array
            items: { type: string }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id:              { type: string }
                  status:              { type: string, enum: [queued, running] }
                  estimated_profiles:  { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /admin/scrape/{job_id}:
    get:
      tags: [admin]
      summary: Get scrape job status
      operationId: getScrapeStatus
      security: [{ adminToken: [] }]
      parameters: [{ name: job_id, in: path, required: true, schema: { type: string }}]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScrapeJobStatus' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/scrape/{job_id}/cancel:
    post:
      tags: [admin]
      summary: Cancel a scrape job
      description: |
        Signals a queued or running job to stop. The job moves to `cancelling` first and to
        `cancelled` once the crawler picks up the signal. Returns 409 if the job already finished.
      operationId: cancelScrape
      security: [{ adminToken: [] }]
      parameters: [{ name: job_id, in: path, required: true, schema: { type: string }}]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScrapeJobStatus' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    adminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
      description: Static admin token. Required for all `/admin/*` routes.
...
  responses:
...
    Unauthorized:
      description: Missing or invalid admin token
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' }}}
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' }}}
    Conflict:
      description: Conflict with current resource state (e.g. cancelling a finished job)
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' }}}
  schemas:
...
    ScrapeJobStatus:
      type: object
      required: [job_id, status, processed, started_at]
      properties:
        job_id:       { type: string }
        status:       { type: string, enum: [queued, running, done, failed, cancelling, cancelled] }
        processed:    { type: integer }
        total:        { type: integer, nullable: true }
        error:        { type: string,  nullable: true }
        started_at:   { type: string,  format: date-time }
        finished_at:  { type: string,  format: date-time, nullable: true }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:    { type: string, example: not_found }
        message: { type: string, example: Person not found }
        detail:  { type: object, additionalProperties: true }
...
security:
  - {}
  - bearerAuth: []